sync composables with pinia store

+++ store/taskStore.ts
@@ -1,6 +1,7 @@
 import { defineStore } from "pinia";
-import { TaskInterface, UpdateTaskInterface } from "@/interfaces/taskInterface"; // Import Task interface from apiFactory
+import { TaskInterface, UpdateTaskInterface } from "@/interfaces/taskInterface";
 import { tasksApiFactory } from "@/apiFactory/task";
+import Swal from "sweetalert2";
 import { useNuxtApp } from "#app";
 
 export const useTaskStore = defineStore("taskStore", {
@@ -16,11 +17,7 @@ export const useTaskStore = defineStore("taskStore", {
         const response = await tasksApiFactory.getAllTasks();
         this.tasks = response.data;
       } catch (error) {
-        const { $toast } = useNuxtApp();
-        $toast.error(error.message, {
-          autoClose: 5000,
-          dangerouslyHTMLString: true,
-        });
+        this.handleError(error);
       } finally {
         this.loading = false;
       }
@@ -31,59 +28,82 @@ export const useTaskStore = defineStore("taskStore", {
         const response = await tasksApiFactory.getTaskById(taskId);
         this.currentTask = response.data;
       } catch (error) {
-        const { $toast } = useNuxtApp();
-        $toast.error(error.message, {
-          autoClose: 5000,
-          dangerouslyHTMLString: true,
-        });
+        this.handleError(error);
       } finally {
         this.loading = false;
       }
     },
     async createTask(task: TaskInterface) {
+      this.loading = true;
       try {
         const response = await tasksApiFactory.createTask(task);
         this.tasks.push(response.data);
-      } catch (error) {
         const { $toast } = useNuxtApp();
-        $toast.error(error.message, {
+        $toast.success('Hurray!!!. New Task was created successfully.', {
           autoClose: 5000,
           dangerouslyHTMLString: true,
         });
+      } catch (error) {
+        this.handleError(error);
       } finally {
         this.loading = false;
       }
     },
     async updateTask(task: UpdateTaskInterface) {
+      console.log(task, "task here");
+      this.loading = true;
       try {
-        const response = await tasksApiFactory.updateTask(task.id, task);
-        const index = this.tasks.findIndex((t) => t.id === task.id);
+        const response = await tasksApiFactory.updateTask(task._id, task);
+        const index = this.tasks.findIndex((t) => t._id === task._id);
         if (index !== -1) {
           this.tasks[index] = response.data;
         }
-      } catch (error) {
         const { $toast } = useNuxtApp();
-        $toast.error(error.message, {
+        $toast.success('Task was updated successfully.', {
           autoClose: 5000,
           dangerouslyHTMLString: true,
         });
+      } catch (error) {
+        this.handleError(error);
       } finally {
         this.loading = false;
       }
     },
     async deleteTask(taskId: number) {
-      try {
-        await tasksApiFactory.deleteTask(taskId);
-        this.tasks = this.tasks.filter((t) => t.id !== taskId);
-      } catch (error) {
-        const { $toast } = useNuxtApp();
-        $toast.error(error.message, {
-          autoClose: 5000,
-          dangerouslyHTMLString: true,
-        });
-      } finally {
-        this.loading = false;
-      }
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!",
+        cancelButtonText: "Nah! Just kidding",
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          this.loading = true;
+          try {
+            await tasksApiFactory.deleteTask(taskId);
+            this.tasks = this.tasks.filter((t) => t._id !== taskId);
+            const { $toast } = useNuxtApp();
+            $toast.success('Task was deleted successfully.', {
+              autoClose: 5000,
+              dangerouslyHTMLString: true,
+            });
+          } catch (error) {
+            this.handleError(error);
+          } finally {
+            this.loading = false;
+          }
+        }
+      });
+    },
+    handleError(error: any) {
+      const { $toast } = useNuxtApp();
+      $toast.error(error.message, {
+        autoClose: 5000,
+        dangerouslyHTMLString: true,
+      });
     },
   },
 });
